+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added hanging client connection release
 - Added `replace` built-in function for mapper scripts 
 - New WebUI features
+- Added `token_ttl_mins` to web_auth config to define auth token expiration duration.
 
 # 3.1.4 (2025-06-17)
 - share live stream refactored